+++ __init__.py
@@ -2,7 +2,7 @@ bl_info = {
     "name": "Reference To Image Plane",
     "description": "Convert a reference image or camera background ref to a textured plane",
     "author": "Samuel Bernou",
-    "version": (0, 3, 2),
+    "version": (0, 4, 0),
     "blender": (2, 93, 0),
     "location": "Object Menu > Convert & Cam Data Properties > Background Images",
     "warning": "",
@@ -12,6 +12,7 @@ bl_info = {
 
 from . import OP_ref_to_image_plane
 from . import ui
+from . import prefs
 from . import OP_cam_bg_img_to_plane
 
 import bpy
@@ -19,6 +20,7 @@ import bpy
 def register():
     if bpy.app.background:
         return
+    prefs.register()
     OP_ref_to_image_plane.register()
     OP_cam_bg_img_to_plane.register()
     ui.register()
@@ -29,6 +31,7 @@ def unregister():
     ui.unregister()
     OP_cam_bg_img_to_plane.unregister()
     OP_ref_to_image_plane.unregister()
+    prefs.unregister()
 
 if __name__ == "__main__":
     register()